Remove np.float references
AttributeError: module 'numpy' has no attribute 'float'.
`np.float` was a deprecated alias for the builtin `float`. To avoid this
error in existing code, use `float` by itself. Doing this will not
modify any behavior and is safe. If you specifically wanted the numpy
scalar type, use `np.float64` here.
The aliases was originally deprecated in NumPy 1.20; for more details
and guidance see the original release note at:
    https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

+++ ehtplot/color/cmap.py
@@ -117,7 +117,7 @@ def ehtrainbow(N=Nq,
                Jp=73.16384, # maximizing minimal Cp for all hue
                Cp=None,
                hp0=32.1526953043875, # offset the hue so that value==0 is red
-               eps=1024*np.finfo(np.float).eps,
+               eps=1024*np.finfo(float).eps,
                **kwargs):
     """Create a perceptually uniform rainbow colormap"""
     name = kwargs.pop('name', "new eht colormap")
@@ -191,7 +191,7 @@ def ehtuniform(N=Nq,
                JpL=6.25,    JpR=93.75, # consistent with 17 quantize levels
                CpL=0.0,     CpR=64.0,
                hpL='coral', hpR='gold', hpD=None,
-               eps=1024*np.finfo(np.float).eps,
+               eps=1024*np.finfo(float).eps,
                **kwargs):
     """Create a perceptually uniform colormap"""
     name = kwargs.pop('name', "new eht colormap")
